+++ wpcusn.php
@@ -3,7 +3,7 @@
  * Plugin Name: WPCUSN - WordPress ClickUp Sync-nator
  * Plugin URI: https://github.com/Krafty-Sprouts-Media-LLC/wpcusn
  * Description: Two-way status synchronization between WordPress and ClickUp
- * Version: 1.3.2
+ * Version: 1.3.3
  * Author: Krafty Sprouts Media, LLC
  * Author URI: https://animalofthings.com
  * GitHub URI: https://github.com/Krafty-Sprouts-Media-LLC/wpcusn
@@ -24,7 +24,7 @@ if (!defined('ABSPATH')) {
 }
 
 // Define plugin constants
-define('WPCUSN_VERSION', '1.3.2');
+define('WPCUSN_VERSION', '1.3.3');
 define('WPCUSN_PLUGIN_DIR', plugin_dir_path(__FILE__));
 define('WPCUSN_PLUGIN_URL', plugin_dir_url(__FILE__));
 define('WPCUSN_PLUGIN_FILE', __FILE__);
@@ -157,4 +157,7 @@ class WPCUSN
 
 // Initialize plugin
 WPCUSN::get_instance();
+
+// Register deactivation hook to clear cron
+register_deactivation_hook( __FILE__, array( 'WPCUSN_Task_Linker', 'clear_cron' ) );